Move weighted_logprob_sum to rl/processors so it is shared across integrations.
The surrogate is not TRL-shaped; keep trl_grpo in integrations/trl and leave ZoRRo/server-side loss opt-in.
